chore: rename *e2e*.js -> *e2e*.mjs
