+++ tests/e2e/yamcs/quickstartSmoke.e2e.spec.mjs
@@ -33,7 +33,8 @@ comfortable running this test during a live mission?" Avoid creating or deleting
 Make no assumptions about the order that elements appear in the DOM.
 */
 
-import { test, expect } from '../opensource/baseFixtures.js';
+import { baseFixtures } from 'openmct-e2e';
+const { test, expect } = baseFixtures;
 
 test.describe("Quickstart smoke tests @yamcs", () => {
     test('Verify that the create button appears and that the Folder Domain Object is available for selection', async ({ page }) => {
