+++ tests/e2e/yamcs/search.e2e.spec.mjs
@@ -24,7 +24,8 @@
 Search Specific Tests
 */
 
-import { test, expect } from '../opensource/pluginFixtures.js';
+import { pluginFixtures } from 'openmct-e2e';
+const { test, expect } = pluginFixtures;
 
 test.describe("Quickstart search tests @yamcs", () => {
     test('Validate aggregate in search result', async ({ page }) => {
